requirements.txt
python-dotenv==1.1.1
python-telegram-bot==22.1
Requests==2.32.4
together==1.5.17
